# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sunstar News API Documentation (सनस्टार न्युज)",
        "description": "Complete RESTful & Real-Time API for Sunstar News Portal: Articles, Live NEPSE & Rashifal Scrapers, Media Uploads/Downloads, Advertisement Banners, Dashboard Stats & Staff Management.",
        "contact": {
            "name": "Sunstar Tech Team",
            "email": "sitaram@sunstarnews.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Laravel Backend API Server"
        },
        {
            "url": "http://localhost:3001",
            "description": "Next.js Frontend Proxy Server"
        }
    ],
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "List all articles with optional filtering",
                "operationId": "93ded9e31119e6bbb22455c72ef97f19",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term in title or content",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of articles (default 20)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Create a new article",
                "operationId": "75bdc5f953261e6f34852386c9b522c0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "नयाँ समाचार शीर्षक"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "मुख्य समाचार"
                                    },
                                    "summary": {
                                        "type": "string",
                                        "example": "समाचार सारांश"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "समाचार विस्तृत सामग्री..."
                                    },
                                    "imageFiles": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "imageUrl": {
                                        "type": "string",
                                        "example": "https://images.unsplash.com/photo-1544735716-392fe2489ffa"
                                    },
                                    "author": {
                                        "type": "string",
                                        "example": "सनस्टार संवाददाता"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Article created successfully"
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get single article by ID or slug",
                "operationId": "a85697b2464e9aaa3f9980a83edbef88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID or slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article details"
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Update existing article",
                "operationId": "beb39fc2f8d02c16b41438aadbd9ba03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Delete article",
                "operationId": "804276a4147e577744e158b2ce2b58f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article deleted successfully"
                    }
                }
            }
        },
        "/api/share": {
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Increment article share counter",
                "operationId": "df112bec6f3641b807a044c97d58074a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "articleId": {
                                        "type": "string",
                                        "example": "art-lead-1"
                                    },
                                    "platform": {
                                        "type": "string",
                                        "example": "facebook"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Share recorded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "shares_count": {
                                            "type": "integer",
                                            "example": 46
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "सेयर संख्या अद्यावधिक भयो!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}